+++ packages/playground/src/api/documents/documentsApi.ts
@@ -7,7 +7,6 @@ import {
   ItemsWrapper,
   MetadataMap,
 } from '@cognite/sdk-core';
-import { PreviewAPI } from './previewApi';
 
 import {
   Document,
@@ -18,11 +17,17 @@ import {
   DocumentsSearchWrapper,
   ExternalDocumentsSearch,
 } from '../../types';
+
+import { PreviewAPI } from './previewApi';
 import { FeedbackAPI } from './feedbackApi';
+import { PipelinesAPI } from './pipelinesApi';
+import { ClassifiersAPI } from './classifiersApi';
 
 export class DocumentsAPI extends BaseResourceAPI<Document> {
   private readonly feedbackAPI: FeedbackAPI;
   private readonly previewAPI: PreviewAPI;
+  private readonly pipelinesAPI: PipelinesAPI;
+  private readonly classifiersAPI: ClassifiersAPI;
 
   constructor(...args: [string, CDFHttpClient, MetadataMap]) {
     super(...args);
@@ -30,6 +35,16 @@ export class DocumentsAPI extends BaseResourceAPI<Document> {
     const [baseUrl, httpClient, map] = args;
     this.previewAPI = new PreviewAPI(baseUrl + '/preview', httpClient, map);
     this.feedbackAPI = new FeedbackAPI(baseUrl + '/feedback', httpClient, map);
+    this.pipelinesAPI = new PipelinesAPI(
+      baseUrl + '/pipelines',
+      httpClient,
+      map
+    );
+    this.classifiersAPI = new ClassifiersAPI(
+      baseUrl + '/classifiers',
+      httpClient,
+      map
+    );
   }
 
   public search = (
@@ -48,9 +63,13 @@ export class DocumentsAPI extends BaseResourceAPI<Document> {
 
   // https://docs.cognite.com/api/playground/#operation/documentsContent
   public content = (
-    ids: DocumentId[]
+    ids: DocumentId[],
+    ignoreUnknownIds?: boolean
   ): Promise<ItemsWrapper<DocumentContent[]>> => {
-    return this.documentContent<ItemsWrapper<DocumentContent[]>>(ids);
+    return this.documentContent<ItemsWrapper<DocumentContent[]>>(
+      ids,
+      ignoreUnknownIds
+    );
   };
 
   public get feedback() {
@@ -59,6 +78,14 @@ export class DocumentsAPI extends BaseResourceAPI<Document> {
 
   public get preview() {
     return this.previewAPI;
+  }
+
+  public get pipelines() {
+    return this.pipelinesAPI;
+  }
+
+  public get classifiers() {
+    return this.classifiersAPI;
   }
 
   private async searchDocuments<ResponseType>(
@@ -72,11 +99,15 @@ export class DocumentsAPI extends BaseResourceAPI<Document> {
   }
 
   private async documentContent<ResponseType>(
-    ids: DocumentId[]
+    ids: DocumentId[],
+    ignoreUnknownIds?: boolean
   ): Promise<ResponseType> {
     const documentIds = ids.map(id => ({ id }));
     const response = await this.post<ResponseType>(this.url('content'), {
-      data: { items: documentIds },
+      data: {
+        items: documentIds,
+        ignoreUnknownIds,
+      },
     });
     return response.data;
   }