+++ packages/playground/src/types.ts
@@ -75,6 +75,27 @@ export interface DocumentsSourceFileFilter {
   size?: Range<number>;
 }
 
+export interface ClassifierName {
+  name?: string;
+}
+
+export interface Classifier extends ClassifierName {
+  id: CogniteInternalId;
+  projectId?: number;
+  createdAt?: number;
+  status?: string;
+  active?: boolean;
+  metrics?: ClassifierMetrics;
+}
+
+export interface ClassifierMetrics {
+  precision?: number;
+  recall?: number;
+  f1Score?: number;
+  confusionMatrix?: number[][];
+  labels?: string[];
+}
+
 export interface ExternalDocumentsSearch {
   filter?: DocumentsFilter;
   search?: DocumentsSearch;
@@ -100,7 +121,7 @@ export interface DocumentsFilter {
   assetIds?: ContainsAllIds | ContainsAnyIds;
   assetSubtreeIds?: ContainsAnyIds | ValueMissing;
   sourceSystem?: StringIn | StringEquals;
-  labels?: Label[];
+  labels?: LabelFilter;
   geoLocation?: GeoLocationFilter;
   sourceFile?: DocumentsSourceFileFilter;
 }
@@ -232,38 +253,93 @@ export interface DocumentsRequestFilter {
   cursor?: string;
 }
 
+export interface Adder<T> {
+  add: T;
+  set?: never;
+  remove?: never;
+}
+
+export interface Remover<T> {
+  remove: T;
+  set?: never;
+  add?: never;
+}
+
+export interface Setter<T> {
+  set: T;
+  remove?: never;
+  add?: never;
+}
+
+export interface NullSetter {
+  setNull: boolean;
+  set?: never;
+  remove?: never;
+  add?: never;
+}
+
+export interface UpdateDocumentsPipeline {
+  externalId: string;
+  sensitivityMatcher?: UpdateDocumentsPipelineSensitivityMatcher;
+  classifier?: UpdateDocumentsPipelineClassifier;
+}
+
+export interface UpdateDocumentsPipelineSensitivityMatcher {
+  matchLists?:
+    | Adder<StringToStringArrayMap>
+    | Setter<StringToStringArrayMap>
+    | Remover<string[]>;
+  fieldMappings?: Setter<DocumentsFieldMappings>;
+  filterPasswords?: Setter<boolean>;
+  sensitiveSecurityCategory?: Setter<boolean> | NullSetter;
+  restrictToSources?: Adder<string[]> | Remover<string[]> | Setter<string[]>;
+}
+
+export interface UpdateDocumentsPipelineClassifier {
+  name?: Setter<string>;
+  trainingLabels:
+    | Adder<LabelList[]>
+    | Remover<LabelList[]>
+    | Setter<LabelList[]>;
+  activeClassifierId: Setter<number> | NullSetter;
+}
+
 export interface DocumentsPipeline {
   externalId: string;
-  sensitivityMatcher?: SensitivityMatcher;
-  classifier?: {
-    trainingLabels: LabelList[];
-  };
+  sensitivityMatcher: SensitivityMatcher;
+  classifier: DocumentsPipelineClassifier;
+}
+
+export interface DocumentsPipelineClassifier {
+  name?: string;
+  trainingLabels: LabelList[];
+  activeClassifierId?: number;
 }
 
 export type LabelList = ExternalId;
 
 export interface SensitivityMatcher {
-  matchLists?: StringToStringArrayMap;
-  fieldMappings?: DocumentsFieldMappings;
+  matchLists: StringToStringArrayMap;
+  fieldMappings: DocumentsFieldMappings;
+  filterPasswords?: boolean;
   sensitiveSecurityCategory?: number;
   restrictToSources?: string[];
 }
 
 export interface DocumentsFieldMappings {
-  title?: string;
-  author?: string;
-  mimeType?: string;
-  type?: string;
+  title?: string[];
+  author?: string[];
+  mimeType?: string[];
+  type?: string[];
   labelsExternalIds?: string[];
   sourceFile?: DocumentsSourceFile;
-  filterPasswords?: boolean;
 }
 
 export interface DocumentsSourceFile {
-  name?: string;
-  directory?: string;
-  content?: string;
-  metadata?: StringToStringMap;
+  name?: string[];
+  directory?: string[];
+  content?: string[];
+  metadata?: StringToStringArrayMap;
 }
 
 export type StringToStringArrayMap = {