+++ packages/stable/src/index.ts
@@ -21,6 +21,7 @@ export {
   OnAuthenticateLoginObject,
   POPUP,
   REDIRECT,
+  CogniteAuthentication,
 } from '@cognite/sdk-core';
 export { default as CogniteClient } from './cogniteClient';