+++ packages/wells/src/client/model/ValueType.ts
@@ -1,5 +1,0 @@
-export const enum ValueType {
-  string = 'STRING',
-  double = 'DOUBLE',
-  long = 'LONG',
-}