+++ packages/wells/src/client/model/WellIds.ts
@@ -1,7 +1,0 @@
-export interface WellIds {
-  /**
-   * @type {number[]}
-   * @memberof WellIds
-   */
-  items?: number[];
-}